Define interface and init function for Request in request.ts
