+++ src/compose.test.ts
@@ -1,5 +1,8 @@
 import { compose } from './compose'
 import { Context } from './context'
+import { extendRequestPrototype } from './request'
+
+extendRequestPrototype()
 
 type C = {
   req: Record<string, string>
